refactor: split customer_service into focused modules (DEBT-1 D2-C)
Mechanical move-only split of customer_service.py (1,088 lines) per the
god-file decomposition plan (docs/plans/2026-06-12-god-file-decomposition.md),
closing the ARCHITECT-001 finding of #428.

- customer_service.py (684): core CRUD, lookup, listing, stats,
  sequence/code generation, pricing. Re-exports the CSV import surface so
  every existing 'from app.services.customer_service import X' and
  'customer_service.X' attribute access keeps working unchanged.
- customer_import_service.py (462, new): CSV parsing, column mapping,
  format detection, preview generation, and bulk import. Code moved verbatim.

No behavior change. ruff clean; full pytest suite green (5539 passed,
13 skipped). The remaining #428 findings (ARCHITECT-002 error handling,
ARCHITECT-003 savepoints, NAVIGATOR-002 preview truncation) are deferred
to D2-D and intentionally untouched here.

Refs #428

Agent-Session: d2c-customer
